Temporarily add `--nogpgcheck` flag to `yum update`
It seems that there is a key issue with the `rockylinux` image builds pertaining to key signing ([build log](https://github.com/rapidsai/ci-imgs/actions/runs/3330586010/jobs/5644101364#step:8:373)).

This PR adds the `--nogpgcheck` flag to get us unblocked for now. We should remove this flag in the future once the key signing issue is resolved.