re-add `timeout-minutes`
I removed this previously because I was under the impression that QEMU jobs will eventually complete even if they take a while.

However, I was proven wrong in the workflow run below, where the workflow timed out after 6 hours due to a QEMU build taking to long.

https://github.com/rapidsai/ci-imgs/actions/runs/3767855767/jobs/6405791187

This commit adds the `timeout-minutes` field back with a value that is generous enough to allow most jobs to complete.