Install AWS CLI from DockerHub Images
This PR changes the way we install the AWS CLI from using `conda`/`curl` to using their official DockerHub images.

There are a few motivations for this change:

- The `awscli` package from `conda-forge` has a lot of dependencies, which can cause packages in the base environment to get downgraded
- The `awscli` package from `conda-forge` has different versions published for `amd64` vs `arm64`, which can cause inconsistencies in our images

The result of this change is that all of our images should include the same AWS CLI version.

I'd like to eventually move more of our dependencies away from using `conda`, but we can save that for another time.