Add `renovate` schedule
This PR configures `renovate` to run once per month on the first day of each month.

This will help eliminate excess GitHub notifications for dependencies that update frequently like the AWS CLI.

Scheduled updates will appear in the dependency dashboard issue below:

- https://github.com/rapidsai/ci-imgs/issues/14

From this dashboard, `renovate` can be triggered to open a PR for a particular dependency outside of the configured schedule.