Drop boa from Conda images (pt. 2)
Upstream has archived `boa` and recommended moving to `rattler-build`.
So this drops `boa` from our images.

Also as `boa` does not support Python 3.13 (unmaintained), dropping
`boa` is a prerequisite to adding Python 3.13. There are other
dependencies it pins to old versions that will be able to get updates
too.

We tried dropping this before, but it caused issues with the RAPIDS
25.02 release. Once the RAPIDS 25.02 release is complete, let's give
this another try.