Use DockerHub for wheel base images too
Currently we use DockerHub for conda base images and `nvcr.io` for wheel
base images. However it looks like updates are not pushed to them at the
same time. As the DockerHub images are already up and have been up for a
little while, use them consistently across both conda and wheel images.